+++ CHANGELOG.md
@@ -10,15 +10,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed  
 
-- Enabled Wyoming region
-
 ### Deprecated 
 
 ### Removed 
 
 ### Fixed  
 
-### Security  
+### Security 
+
+## [v4.16.0](https://github.com/USGS-WiM/StreamStats/releases/tag/v4.16.0) - 2023-06-22
+
+### Changed  
+
+- Enabled Wyoming region
 
 ## [v4.15.0](https://github.com/USGS-WiM/StreamStats/releases/tag/v4.15.0) - 2023-05-24